# lib/api-zod/src/generated/types/gameSessionEventInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Souqrates System API
 * OpenAPI spec version: 0.1.0
 */

export interface GameSessionEventInput {
  /** Whether this hit was correct (scores) or wrong (penalty). */
  correct: boolean;
  /** Base points the engine awarded for a correct hit (pre-combo). The server clamps this to a safe range; ignored for wrong hits. */
  points?: number;
}