# lib/api-zod/src/generated/types/gameSessionEventResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Souqrates System API
 * OpenAPI spec version: 0.1.0
 */

export interface GameSessionEventResult {
  score: number;
  combo: number;
  reachedTarget: boolean;
  throttled?: boolean;
}